+++ contracts/Interfaces/IPriceFeed.sol
@@ -3,5 +3,5 @@
 pragma solidity 0.7.6;
 
 interface IPriceFeed {
-    function fetchPrice() external returns (uint);
+    function fetchPrice() external returns (uint256);
 }
